Electron web frame zoom factor update & elementDefinition parser has been moved to utils
