+++ src/common/services/fhir.service.ts
@@ -2,7 +2,6 @@ import { environment } from '../environment'
 import { FhirClient } from 'ng-fhir/FhirClient'
 import axios from 'axios'
 import http from 'http'
-import electronStore from './../electron-store'
 
 export class FhirService {
 
@@ -153,86 +152,6 @@ export class FhirService {
    */
   deleteAll (resourceType: string) {
     return axios.delete(this.config.baseUrl + '/' + resourceType)
-  }
-
-  /**
-   * Parses elements of a StructureDefinition resource (StructureDefinition.snapshot.element)
-   * @param parameter - Search parameter
-   * @param profileId
-   */
-  parseElementDefinitions (parameter: string, profileId: string): Promise<any> {
-    const fhirService: FhirService = new FhirService(electronStore.get('fhirBase'))
-    const query = {}
-    query[parameter] = profileId
-    return new Promise((resolveParam, rejectParam) => {
-      fhirService.search('StructureDefinition', query, true)
-        .then(res => {
-          const bundle = res.data as fhir.Bundle
-          if (bundle.entry?.length) {
-            const resource = bundle.entry[0].resource as fhir.StructureDefinition
-            const list: fhir.ElementTree[] = []
-            Promise.all(resource?.snapshot?.element.map((element: fhir.ElementDefinition) => {
-              return new Promise(resolveElement => {
-                const parts = element?.id?.split('.') || []
-                let tmpList = list
-                Promise.all(parts.map(part => {
-                  return new Promise((resolveElementPart => {
-                    let match = tmpList.findIndex(l => l.label === part)
-                    if (match === -1) {
-                      match = 0
-                      const item: fhir.ElementTree = {
-                        value: element?.id,
-                        label: part,
-                        definition: element?.definition,
-                        comment: element?.comment,
-                        short: element?.short,
-                        min: element?.min,
-                        max: element?.max,
-                        type: [],
-                        children: []
-                      }
-                      Promise.all(element.type?.map((_: fhir.ElementDefinitionType) => {
-                        return new Promise(resolveElementType => {
-                          if (_.code && _.code[0] === _.code[0].toUpperCase() && environment.datatypes[_.code]) {
-
-                            const cached = electronStore.get(`datatype-${_.code}`)
-                            if (cached) {
-                              item.type?.push(cached)
-                              resolveElementType()
-                            } else {
-                              if (_.code === 'Reference') resolveElementType()
-                              this.parseElementDefinitions('url', environment.datatypes[_.code])
-                                .then((elementTreeList: fhir.ElementTree[]) => {
-                                  elementTreeList.length ? item.type?.push({...elementTreeList[0]}) : {}
-                                  electronStore.set(`datatype-${_.code}`, {...elementTreeList[0]})
-                                  resolveElementType()
-                                })
-                                .catch(err => resolveElementType())
-                            }
-                          } else {
-                            item.type?.push({value: _.code})
-                            resolveElementType()
-                          }
-                        })
-                      }) || []).then(() => resolveElementPart()).catch(() => resolveElementPart())
-                      tmpList.push(item)
-                    } else resolveElementPart()
-                    tmpList = tmpList[match].children as fhir.ElementTree[]
-                  }))
-                })).then(() => resolveElement()).catch(() => resolveElement())
-              })
-            }) || [])
-              .then(() => resolveParam(list))
-              .catch(() => rejectParam([]))
-          } else {
-            resolveParam([])
-          }
-        })
-        .catch(err => {
-          rejectParam([])
-        })
-    })
-
   }
 
 }
