+++ src/main.ts
@@ -5,6 +5,9 @@ import store from './store'
 import './quasar'
 import { QVueGlobals } from 'quasar'
 import logger from './common/logger'
+import { webFrame } from 'electron'
+
+webFrame.setZoomFactor(0.9)
 
 Vue.config.productionTip = false
